Remove Old Encryption Code
The file is badly outdated and relies on a library that could requires
the user to only use characters specified within certain bounds.
The code is poor and should be not reused.

Note: A new library for providing encryption capabilities can be found
under the lib folder. "encrypt.lua" is not finished but is making headway
much faster than "encode.lua" All further development on encryption will
shift over.